Rename wire_types to bam_types
