+++ application/comit_node/src/comit_client/mod.rs
@@ -5,14 +5,14 @@ use futures::Future;
 use std::{io, net::SocketAddr, sync::Arc};
 
 use std::{fmt::Debug, panic::RefUnwindSafe};
-use swap_protocols::{rfc003, wire_types};
+use swap_protocols::{bam_types, rfc003};
 
 pub trait Client: Send + Sync {
     fn send_swap_request<
         SL: rfc003::Ledger,
         TL: rfc003::Ledger,
-        SA: Into<wire_types::Asset>,
-        TA: Into<wire_types::Asset>,
+        SA: Into<bam_types::Asset>,
+        TA: Into<bam_types::Asset>,
     >(
         &self,
         request: rfc003::Request<SL, TL, SA, TA>,
