+++ application/comit_node/src/swap_protocols/asset.rs
@@ -1,10 +1,10 @@
 use bitcoin_support::BitcoinQuantity;
 use ethereum_support::EtherQuantity;
 use std::fmt::Debug;
-use swap_protocols::wire_types;
+use swap_protocols::bam_types;
 
 pub trait Asset:
-    Clone + Debug + Send + Sync + 'static + PartialEq + Into<wire_types::Asset>
+    Clone + Debug + Send + Sync + 'static + PartialEq + Into<bam_types::Asset>
 {
 }
