+++ application/comit_node/src/swap_protocols/handler.rs
@@ -1,4 +1,4 @@
-use swap_protocols::wire_types::SwapResponse;
+use swap_protocols::bam_types::SwapResponse;
 
 pub trait SwapRequestHandler<Req>: Send + 'static {
     fn handle(&mut self, _request: Req) -> SwapResponse {
